fixed card widths & removed template code from containers
