+++ src/components/Sidebar/Sidebar.tsx
@@ -3,6 +3,22 @@ import { Link } from "react-router-dom";
 import placeholder from "../../placeholder.png";
 import "./Sidebar.css";
 
+const sidebarItem = (props: { to: string; text: string; imageUrl?: string }) => {
+  const { to, text, imageUrl } = props;
+  const imageSrc: string = imageUrl || placeholder;
+
+  return (
+    <li className="padding-v-normal">
+      <Link to={to}>
+        <img src={imageSrc} height="48" />
+        <div className="type-small">
+          <span>{text}</span>
+        </div>
+      </Link>
+    </li>
+  );
+};
+
 class Sidebar extends React.Component {
   public render() {
     return (
@@ -16,24 +32,10 @@ class Sidebar extends React.Component {
             <img src={placeholder} height="48" />
             <div className="type-small">Functions</div>
           </li>
-          <li className="padding-v-normal">
-            <img src={placeholder} height="48" />
-            <div className="type-small">
-              <Link to={"/charts"}>Charts</Link>
-            </div>
-          </li>
-          <li className="padding-v-normal">
-            <img src={placeholder} height="48" />
-            <div className="type-small">
-              <Link to={"/services"}>Service Catalog</Link>
-            </div>
-          </li>
-          <li className="padding-v-normal">
-            <img src={placeholder} height="48" />
-            <div className="type-small">
-              <Link to={"/repos"}>App Repositories</Link>
-            </div>
-          </li>
+          {sidebarItem({ to: "/charts", text: "Charts" })}
+          {sidebarItem({ to: "/charts", text: "Charts" })}
+          {sidebarItem({ to: "/services", text: "Service Catalog" })}
+          {sidebarItem({ to: "/repos", text: "App Repositories" })}
         </ul>
       </aside>
     );
